requirements.txt
flask
flask-cors
flask-limiter
flask-talisman
supabase
requests
python-dotenv
gunicorn
